# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fikrie Portfolio API
  description: |
    REST API for the Fikrie Portfolio Next.js application.

    ## Authentication
    Most write endpoints require a session cookie:
    - Cookie name: `portfolio_session`
    - Cookie value: `authenticated`

    Log in via the `/login` page (Server Action, not a REST endpoint) to obtain the cookie,
    then include it in subsequent requests.

    ## Data storage
    - JSON data: `data/*.json`
    - Uploaded files: `public/` subdirectories
  version: 1.0.0
  contact:
    name: Fahreza Portfolio
servers:
  - url: http://localhost:3000
    description: Local development
  - url: http://localhost:3001
    description: Playwright / alternate dev port
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: portfolio_session
      description: Set to `authenticated` after logging in
  schemas:
...
    PortfolioClient:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
      required: [id, name, description]
...
    AccountInfo:
      type: object
      properties:
        username: { type: string }
        recoveryEmail: { type: string }
        createdAt: { type: string, format: date-time }
paths:
...
  /api/contact:
    post:
      tags: [Contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2 }
                email: { type: string, format: email }
                subject: { type: string, minLength: 2 }
                message: { type: string, minLength: 10 }
                website: { type: string, description: Honeypot field }
              required: [name, email, subject, message]
      responses:
        '200':
          description: Message saved; may return clientEmailDelivery for FormSubmit
        '400': { description: Validation error }
        '503': { description: Email not configured }
...
  /api/portfolio/{sectionId}:
    parameters:
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          enum: [syncsoft, telstra-thealth, enett, dws, kmart]
    patch:
      tags: [Portfolio]
      summary: Update portfolio section text
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                period: { type: string }
                company: { type: string, minLength: 2 }
                role: { type: string }
                description: { type: string, minLength: 10 }
                clients:
                  type: array
                  items: { $ref: '#/components/schemas/PortfolioClient' }
              required: [company, description]
      responses:
        '200':
          description: Returns section and sections array
        '401': { description: Unauthorized }
        '400': { description: Invalid section or validation error }
  /api/portfolio/{sectionId}/media:
    parameters:
      - name: sectionId
        in: path
        required: true
        schema:
          type: string
          enum: [syncsoft, telstra-thealth, enett, dws, kmart]
    post:
      tags: [Portfolio]
      summary: Upload portfolio section image
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                caption: { type: string }
              required: [file]
      responses:
        '200': { description: Returns image and updated section }
        '401': { description: Unauthorized }
    delete:
      tags: [Portfolio]
      summary: Delete portfolio section image
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageId: { type: string }
              required: [imageId]
      responses:
        '200': { description: Returns updated section }
        '401': { description: Unauthorized }
...
  /api/settings/account:
    get:
      tags: [Settings]
      summary: Get admin account info
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  account: { $ref: '#/components/schemas/AccountInfo' }
        '401': { description: Unauthorized }
    patch:
      tags: [Settings]
      summary: Update account (username, recovery email, password)
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                recoveryEmail: { type: string, format: email }
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 6 }
      responses:
        '200': { description: Returns updated account }
        '401': { description: Unauthorized }
        '400': { description: Validation error }
  /api/settings/profile-image:
    post:
      tags: [Settings]
      summary: Upload hero profile image
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
              required: [file]
      responses:
        '200': { description: Returns profileImage path and updated content }
        '401': { description: Unauthorized }
...